Create streamlit.time_utils module and move functions into it
