+++ lib/streamlit/elements/widgets/time_widgets.py
@@ -50,6 +50,7 @@ from streamlit.runtime.state import (
     register_widget,
 )
 from streamlit.runtime.state.common import compute_widget_id
+from streamlit.time_util import adjust_years
 from streamlit.type_util import Key, LabelVisibility, maybe_raise_label_warnings, to_key
 
 if TYPE_CHECKING:
@@ -65,23 +66,6 @@ DEFAULT_STEP_MINUTES: Final = 15
 ALLOWED_DATE_FORMATS: Final = re.compile(
     r"^(YYYY[/.\-]MM[/.\-]DD|DD[/.\-]MM[/.\-]YYYY|MM[/.\-]DD[/.\-]YYYY)$"
 )
-
-
-def _adjust_years(input_date: date, years: int) -> date:
-    """Add or subtract years from a date."""
-    try:
-        # Attempt to directly add/subtract years
-        return input_date.replace(year=input_date.year + years)
-    except ValueError as err:
-        # Handle case for leap year date (February 29) that doesn't exist in the target year
-        # by moving the date to February 28
-        if input_date.month == 2 and input_date.day == 29:
-            return input_date.replace(year=input_date.year + years, month=2, day=28)
-
-        raise StreamlitAPIException(
-            f"Date {input_date} does not exist in the target year {input_date.year + years}. "
-            "This should never happen. Please report this bug."
-        ) from err
 
 
 def _parse_date_value(
@@ -128,9 +112,9 @@ def _parse_min_date(
         parsed_min_date = min_value
     elif min_value is None:
         if parsed_dates:
-            parsed_min_date = _adjust_years(parsed_dates[0], years=-10)
+            parsed_min_date = adjust_years(parsed_dates[0], years=-10)
         else:
-            parsed_min_date = _adjust_years(date.today(), years=-10)
+            parsed_min_date = adjust_years(date.today(), years=-10)
     else:
         raise StreamlitAPIException(
             "DateInput min should either be a date/datetime or None"
@@ -149,9 +133,9 @@ def _parse_max_date(
         parsed_max_date = max_value
     elif max_value is None:
         if parsed_dates:
-            parsed_max_date = _adjust_years(parsed_dates[-1], years=10)
+            parsed_max_date = adjust_years(parsed_dates[-1], years=10)
         else:
-            parsed_max_date = _adjust_years(date.today(), years=10)
+            parsed_max_date = adjust_years(date.today(), years=10)
     else:
         raise StreamlitAPIException(
             "DateInput max should either be a date/datetime or None"
