+++ lib/streamlit/runtime/caching/cache_resource_api.py
@@ -46,7 +46,8 @@ from streamlit.runtime.caching.hashing import HashFuncsDict
 from streamlit.runtime.metrics_util import gather_metrics
 from streamlit.runtime.scriptrunner.script_run_context import get_script_run_ctx
 from streamlit.runtime.stats import CacheStat, CacheStatsProvider, group_stats
-from streamlit.util import TimedCleanupCache, time_to_seconds
+from streamlit.time_util import time_to_seconds
+from streamlit.util import TimedCleanupCache
 
 _LOGGER: Final = get_logger(__name__)
