+++ lib/streamlit/runtime/fragment.py
@@ -29,7 +29,7 @@ from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
 from streamlit.runtime.metrics_util import gather_metrics
 from streamlit.runtime.scriptrunner import get_script_run_ctx
 from streamlit.runtime.scriptrunner.script_run_context import dg_stack
-from streamlit.util import time_to_seconds
+from streamlit.time_util import time_to_seconds
 
 F = TypeVar("F", bound=Callable[..., Any])
 Fragment = Callable[[], Any]
