+++ lib/tests/streamlit/elements/time_widgets_test.py
@@ -17,7 +17,6 @@ import datetime
 from parameterized import parameterized
 
 import streamlit as st
-from streamlit.elements.widgets.time_widgets import _adjust_years
 from streamlit.errors import StreamlitAPIException
 from tests.delta_generator_test_case import DeltaGeneratorTestCase
 
@@ -58,30 +57,3 @@ class StreamlitAPITest(DeltaGeneratorTestCase):
             st.time_input("Set an alarm for", value, step=datetime.timedelta(hours=24))
         with self.assertRaises(StreamlitAPIException):
             st.time_input("Set an alarm for", value, step=datetime.timedelta(days=1))
-
-
-@parameterized.expand(
-    [
-        # Test adding one year to a regular date:
-        (datetime.date(2021, 1, 1), 1, datetime.date(2022, 1, 1)),
-        # Test adding one year to a leap day, resulting in a non-leap year (Feb 28):
-        (datetime.date(2020, 2, 29), 1, datetime.date(2021, 2, 28)),
-        # Test subtracting one year from a leap day, adjusting to the previous year's Feb 28.
-        (datetime.date(2020, 2, 29), -1, datetime.date(2019, 2, 28)),
-        # Non-leap year, regular date
-        (datetime.date(2021, 3, 15), 1, datetime.date(2022, 3, 15)),
-        # Leap year to leap year:
-        (datetime.date(2020, 2, 29), 4, datetime.date(2024, 2, 29)),
-        # Subtracting years from a non-leap year date
-        (datetime.date(2019, 3, 15), -1, datetime.date(2018, 3, 15)),
-        # Adding 100 years, including a century leap year:
-        (datetime.date(1920, 4, 10), 100, datetime.date(2020, 4, 10)),
-        # End of year date
-        (datetime.date(2019, 12, 31), 1, datetime.date(2020, 12, 31)),
-    ]
-)
-def test_adjust_years(
-    input_date: datetime.date, years: int, expected_date: datetime.date
-):
-    """Test that `_adjust_years` correctly` adjusts the year of a date."""
-    assert _adjust_years(input_date, years) == expected_date
