+++ lib/tests/streamlit/util_test.py
@@ -14,11 +14,9 @@
 
 import asyncio
 import gc
-import math
 import random
 import unittest
-from datetime import timedelta
-from typing import Any, Dict, List, Set
+from typing import Dict, List, Set
 from unittest.mock import patch
 
 import numpy as np
@@ -26,19 +24,6 @@ import pytest
 from parameterized import parameterized
 
 from streamlit import util
-
-TIME_STRING_TO_SECONDS_PARAMS = [
-    ("float", 3.5, 3.5),
-    ("timedelta", timedelta(minutes=3), 60 * 3),
-    ("str 1 arg", "1d", 24 * 60 * 60),
-    ("str 2 args", "1d23h", 24 * 60 * 60 + 23 * 60 * 60),
-    (
-        "complex str 3 args",
-        "1 day 23hr 45minutes",
-        24 * 60 * 60 + 23 * 60 * 60 + 45 * 60,
-    ),
-    ("str 2 args with float", "1.5d23.5h", 1.5 * 24 * 60 * 60 + 23.5 * 60 * 60),
-]
 
 
 class UtilTest(unittest.TestCase):
@@ -223,37 +208,3 @@ class UtilTest(unittest.TestCase):
             assert len(asyncio.all_tasks()) == 1
 
         asyncio.run(check())
-
-    @parameterized.expand(
-        [
-            *TIME_STRING_TO_SECONDS_PARAMS,
-            ("None", None, math.inf),
-        ]
-    )
-    def test_time_to_seconds_coerced(
-        self, _, input_value: Any, expected_seconds: float
-    ):
-        """Test the various types of input that time_to_seconds accepts."""
-        assert expected_seconds == util.time_to_seconds(input_value)
-
-    @parameterized.expand(
-        [
-            *TIME_STRING_TO_SECONDS_PARAMS,
-            ("None", None, None),
-        ]
-    )
-    def test_time_to_seconds_not_coerced(
-        self, _, input_value: Any, expected_seconds: float
-    ):
-        """Test the various types of input that time_to_seconds accepts."""
-        assert expected_seconds == util.time_to_seconds(
-            input_value, coerce_none_to_inf=False
-        )
-
-    def test_time_str_exception(self):
-        """Test that a badly-formatted time string raises an exception."""
-        with pytest.raises(util.BadTimeStringError):
-            util.time_to_seconds("")
-
-        with pytest.raises(util.BadTimeStringError):
-            util.time_to_seconds("1 flecond")
